feat: parallel count tasks on series_id
